cpu: remove di/dimux latch
Once all state is guarded by RDY, it is no longer necessary to latch DI.
Since there are states, unfortunately, where core outputs are combinatorial
functions of inputs, it means that outputs may not be stable during ~RDY,
notably AB, DO, and WE.

If the user keeps core inputs stable during ~RDY, the outputs will also be
stable. This will normally be the case for synchronous memory. However, be
aware that IRQ and NMI also have combinatorial paths back to the core outputs.

This is still always correct behavior; it is the users responsibility not to
latch or depend on any outputs unless RDY is true. If external memory requires
stable AB, the user may add a latch or register externally.

equiv: affects flags, AI, BI, IR, AB, PC_temp which were not guarded by RDY